extract common function in service, ingress and federation e2e to util, to be reused.
